requirements.txt
Django==4.2
djangorestframework==3.14.0
requests==2.28.2
python-dotenv
